Change Linux to report modify when close_write is detected.
The MODIFY event is fired any time a write syscall is fired.

This means that it is often received while a file is being saved, and/or sent
multiple times.

This change makes it only report the close of a file opened for writing,
which means only one event is received, and the writes have all completed
before the event is sent.