Fix some warnings in tests/notify.rs
std::thread::sleep_ms (deprecated) was replaced with called to
std::thread::sleep. A couple of other warnings were fixed as well.

There's still a number of items in this file with dead_code warnings
that could be fixed by running the tests on all platforms, but from the
comments, it sounds like some of the tests are flakey in certain
scenarios.