Rename and rescope polling backend
I was being inordinately silly: one entire point of this whole v5 thing 
is to reduce duplication. Processors that fill in capabilities are 
already going to have to read directories and recurse, so the base 
poller really only needs to be a *file* poller. Not a folder tree 
poller!

That should make it a whole lot simpler to implement, as well as a good 
target for implementing the required processors.