[tests] Translate debounced events
In a way, this commit is discarding a lot of information by only 
testing, with one exception, the top-level kind of the event, instead of 
the precise event like Op::METADATA or Op::WRITE (or equivalents).

However, that loss is actually the point: Notify is not in the business 
of presenting precise events consistently across platforms, it's in the 
business of presenting a _simplified_ model of events as consistently as 
it can, while also exposing the precise nature of the events if that's 
needed.

While this commit busies itself with simply converting the tests to the 
new event format, it leaves a lot of low-hanging fruit to remove the 
platform-based distinctions in the tests, and therefore bring us closer 
to a world that is superficially consistent.