[docs] Push minimum rustc to 1.32.0
The ecosystem is moving to edition 2018 wholesale and bringing a lot of
breaking changes via dependency versioning for crates that aren't
compatible. I don't like adding yet one more change to this release, but
the plan is NOT to bring the code to 2018 standard, instead to push the
CI and guarantees and leave everything as is for the first release.
Later point releases will be able to refactor for new features and into
the new edition.