Improve fsevent thread safety
While I couldn't prove any concrete issues before this patch, it can be risky
to have both the event stream thread and the main thread access the context
at the same time, since it could lead to data races or undefined behavior.
This changes the handling of the context to factor out the `done` channel
from the context, and to move explicitly dropping the context once the
stream has been stopped.

Furthermore, this creates some helper wrappers to send CFRef types across
threads instead of casting values to integers.