Make EventFn take FnMut
This changes EventFn to work with FnMut, rather than Fn. The advantage of
this is that this allows callers to use state in the EventFn, rather than
having to use interior mutability. This is helpful particularly for
getting notify to work with async rust, since `futures::channel::mpsc`'s
methods all take `&mut self`.

The downside of this change is that it limits the ability of notify to
share `EventFn` across multiple threads, but all the backends either
guarantee only a single background thread, or they already use an
`Arc<Mutex<...>>` to protect `EventFn`.