Make `Watcher` object safe
This changes the `Watcher` trait to be object safe, which allows users to
dynamically select which backend to use. This can be helpful on systems that
support multiple file watching systems that have different tradeoffs. For
example, Chromium's file watcher on OS X will use an fsevent backend
when watching recursive directories, and a kqueues backend when not.

In order to implement this, this makes a few changes to Watcher:

* removes `new_immediate` constructor.
* removes the `<P: AsRef<Path>>` from the `watch` and `unwatch`.

This then replaces (and renames) calls with `notify::recommended_watcher()`
to get similar behavior as the old constructor.