kqueue: watch each file not just the directories
The kqueue backend started as a copy of the inotify backend.
One major difference between kqueue and inotify was overlook by me,
inotify reports events for changes to files contained in a folder if the
folder is watched. Kqueue dosen't do this, each file has to be watch
individually.

This change fixes this, events are now reported for all files in a
recursive watch.