inotify: Ignore any error sending the `RenameTimeout` event.
The `unwrap()` call here causes thread panics in our application. We notice
them because we use `set_panic_hook`. I think ordinarily a panic here would be
printed to stderr and otherwise ignored.

For us, I think the usual pattern is that several files are renamed at once,
and as soon as we notice the first rename, we drop the watcher. Therefore, when
this thread wakes up, it's trying to send an event to a queue that has been
closed.