Do not remove a filename twice
When unwatching a directory that was watched recursively, notify's kqueue impl
would first unwatch the containing directory, then walk it to unwatch all
the items inside.

However, walking the directory first processes the containing directory, and
unwatching a directory twice can cause an error in the kqueue crate.

Fixes #665 and adds a test to prevent regressions.