fix: INotifyWatcher may raise events with no paths (#700)
* fix: INotifyWatcher may raise events with no paths

INotifyWatcher hold inotify watch descriptors map `wd -> path` and after an `unwatch` call  the descriptor is deleted from the map.
In some cases, if a lot of events are raised, it may cause race condition when the descriptor has been deleted, but event with
this descriptor has already been queued. In that case INotifyWatcher doesn't know the path and just ignore the event

fixes #678

* chore: changelog for (#700)

* fix: typo

* fix: typo

Co-authored-by: Daniel Faust <hessijames@mailbox.org>

---------

Co-authored-by: Daniel Faust <hessijames@mailbox.org>